change confirmations type from uint8 to uint128

+++ contracts/interfaces/ITrustedGuardiansPlugin.sol
@@ -11,14 +11,14 @@ interface ITrustedGuardiansPlugin {
     event SetGuardian(bytes32 indexed guardian, bool indexed _isAdd);
     event GuardiansVoteForDistribution(bytes32 indexed guardian, uint256 votedCount);
     event GuardiansDistributionStartSet(bytes32 indexed guardian, uint256 distributionStartAt);
-    event SetGuardiansConfig(uint8 guardiansThreshold, uint64 guardiansChallengeTimeout);
+    event SetGuardiansConfig(uint128 guardiansThreshold, uint64 guardiansChallengeTimeout);
     event ResetGuardiansVoting();
     event ClearGuardiansVoted();
 
     struct PluginStorage {
         EnumerableSet.Bytes32Set guardians;
         bytes32[] guardiansVoted;
-        uint8 guardiansThreshold;
+        uint128 guardiansThreshold;
         uint64 guardiansChallengeTimeout;
     }
     struct GuardianToChange {
